requirements.txt
asgiref==3.9.1
coverage==6.5.0
dj-database-url==3.0.1
dj-static==0.0.6
Django==5.2.5
django-modeltranslation==0.18.5
django-stdimage==6.0.1
gunicorn==20.1.0
model-mommy==2.0.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python-decouple==3.8
sqlparse==0.4.2
static3==0.7.0
typing_extensions==4.14.1
